split read preference tests
